FEATURE: wui: Add clickable graph icon to the flags column
A click on the icon opens a popup that points to the following URL:
/monitoring-graph?device=%device%&indicator=%indicator%

The %device% and %indicator% placeholders gets replaced respectively
by the Nagios host and service names. For host rows, the %indicator%
placeholder gets replaced by the "hostcheck" string.

URL rewrite can be done on the server side to redirect the client to
the graph vizualization tool integrated on the monitoring platform,
for instance a Grafana dashboard that understands device/indicator
passed as query string parameters in the URL.