MAJOR: Move API config to JSON and make API queries dynamic
The API config was defined in "main.js" a Vue instance prototype,
meaning it was only available in a Vue instance and only made available
after the build of "main.js". This resulted in the config not being
available in ".js" files, and also not being availablein components upon
startup (until main.js is built) which creates limitations for future
updates. It was moved to a static config JSON file that is available
everywhere as an import.

After the changes done to the configuration, some changes were needed to
make use of the API URL set up in the config file. Queries to the API
are now dynamic, meaning that changing the API base URL in the config
file will automatically extend to all queries within the App.