REORG/MAJOR: panel: Integrate AutoTable component to render the table
The AutoTable component, that was imported to this project, was added to
the Panel component in order to replace the current data-table in place.
Some logic regarding headers configuration was modified so that the
headers could be used by the AutoTable.

This is a work in progress, and some formatting of the cells and headers
still needs to be converted / added back into the logic.