REORG/MAJOR: apis: Externalize API data fetching calls
In order to handle the fetching of data from an API in a dedicated
plugin, the "api-caller" plugin was created. It is now its responsibility
to call APIs using Axios in order to fetch data, and to feed it to the
components where it is called (Panel for instance). The Panel now only
handles the data response from the plugin and works with it to generate
the table data needed by the app.