MAJOR: Remove all direct imports of the config file
In an effort to remove all direct imports of the config.json file across
the whole app, some changes were made regarding api related info, as it
used to be retrieved directly from the config file.

Now, the app on startup will fetch the config file without compiling it,
and will only store necessary information (base api urls, api type, data
path, etc.) in a config object made available by the api-manager.

This config object is then being watch by the components that depend on
its data. Upon receiving a change in the config object (meaning the
config has been successfully been fetched) their different handlers are
triggered.

In API specific managers, "queryUrls" was turned into a function because
when it was a variable it was evaluated on startup, when api data had
not yet been received. Now the function is only called once the api
config has been fetched.