MAJOR: config: Load config early
Before this commit the config was loaded asynchronously in the main App
component. As a result, handling the apiType readiness correctly was too
much hassle. Everything is much simpler now, reducing complexity in
components that need access to apiConfig (apiType). The config is fetched
first, synchronously using an await, before instantiating Vue.