MINOR: npm: Set proper name and url, remove version
This is not a published npm package, so we can remove the version
field, we would not maintain it in package.json.

As for the change in vue.config.js, nodejs does string coercion
when assigning an undefined value to a process.env property. This
is because environment variable are always strings by nature.
As it was, we would endup with string 'undefined' in
process.env.VUE_APP_VERSION. The patch won't assign process.env
properties if the value would be undefined, so now we get a real
undefined value when accessing process.env.VUE_APP_VERSION from
the app, which Vue will render as an empty string in the sidebar
(App.vue).