FEAT: Add V1 of log viewer as "History"
A new Feature was added to Kompot : the History view.
It's a new view that's intended for viewing logs of indicators & hosts report states between two separate dates/times. Currently the 2 dates are arbitrary in the config but will be changeable in the future.

This view can be accessed via its new entry in the navigation menu, and simply features an AutoTable : it is entirely set up via its config JSON which is currently fetched from the local kompot config file, but will be moved to a remote API at a later date.