BUG/MINOR: menu: The edit icon hover state is per a menu item
As a rule of thumb, avoid mouseenter/mouseleave JS events for style
change on a mouse hover state, use CSS instead when possible. But
here the main problem was with the <hovered> state variable because
it is global and applies to all menu items. Thus if the menu as more
than one drawio item, hovering the edit icon on one item would change
the color for all edit icons.