MINOR: app: Listen for events from other Kompot browsing contexts
Use the broadcast channel API to exchage messages between windows, tabs,
frames, ... related to Kompot. Here we are going to use it to receive
messages from the drawio editor when a diagram get saved. The idea is
to reload the diagram in Kompot viewer if we get notified of a change
made from the editor.

Broadcast Channel API
https://developer.mozilla.org/en-US/docs/Web/API/Broadcast_Channel_API