FEATURE: drawio: New viewer component for a better integration in Vue
The diagram name to load is now passed via route params
/drawio/view/:schema. That component is still based on an iframe,
but here the src attribute is computed using the configured
drawioViewUrl and the schema route parameter.

The component also listens for kchan.drawio.saved events from the
internal event bus to trigger a reload of the iframe when the
schema changes.